App.jsx to HomePage.jsx
